# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/requests": {
      "get": {
        "operationId": "RequestsController_list",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowState",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "approverId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentApproverId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "substituteId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      },
      "post": {
        "operationId": "RequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}": {
      "get": {
        "operationId": "RequestsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/events": {
      "get": {
        "operationId": "RequestsController_events",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/vacation": {
      "post": {
        "operationId": "RequestsController_createVacation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVacationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/approve": {
      "post": {
        "operationId": "RequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/reject": {
      "post": {
        "operationId": "RequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/manager-approve": {
      "post": {
        "operationId": "RequestsController_managerApprove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerApproveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/manager-reject": {
      "post": {
        "operationId": "RequestsController_managerReject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/hr-confirm": {
      "post": {
        "operationId": "RequestsController_hrConfirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/hr-reject": {
      "post": {
        "operationId": "RequestsController_hrReject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithRequiredNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/substitute/accept": {
      "post": {
        "operationId": "RequestsController_substituteAccept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/substitute/decline": {
      "post": {
        "operationId": "RequestsController_substituteDecline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithRequiredNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/return": {
      "post": {
        "operationId": "RequestsController_returnForRevision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithRequiredNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/{id}/cancel": {
      "post": {
        "operationId": "RequestsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/bulk-approve": {
      "post": {
        "operationId": "RequestsController_bulkApprove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkApproveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
    "/api/requests/bulk-reject": {
      "post": {
        "operationId": "RequestsController_bulkReject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRejectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["requests"]
      }
    },
...
  },
  "info": {
    "title": "OpenClockwork API",
    "description": "Self-hostable working-time tracker — REST + WebSocket surface.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["Vacation", "HomeOffice", "SpecialLeave", "TimeAdjustment"]
          },
          "from": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-03T00:00:00.000Z"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-07T00:00:00.000Z"
          },
          "reason": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "required": ["employeeId", "type", "from", "to"]
      },
      "CreateVacationDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "substituteId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "reason": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          },
          "halfDayStart": {
            "type": "boolean",
            "default": false,
            "description": "Take the first day as a half-day"
          },
          "halfDayEnd": {
            "type": "boolean",
            "default": false,
            "description": "Take the last day as a half-day"
          }
        },
        "required": ["employeeId", "from", "to"]
      },
      "TransitionDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "required": ["actorId"]
      },
      "ManagerApproveDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          },
          "requiresHrConfirmation": {
            "type": "boolean",
            "default": false
          }
        },
        "required": ["actorId"]
      },
      "TransitionWithRequiredNoteDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": ["actorId", "note"]
      },
      "BulkApproveDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "note": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          },
          "requiresHrConfirmation": {
            "type": "boolean",
            "default": false
          }
        },
        "required": ["actorId", "ids"]
      },
      "BulkRejectDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "note": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": ["actorId", "ids", "note"]
      },
...
    }
  }
}